Now have techlib file
